requirements.txt
asgiref==3.7.2
astroid==2.15.5
attrs==23.1.0
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
flake8==6.0.0
flake8-bugbear==23.6.5
flake8-coding==1.3.2
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg==3.1.9
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pylint==2.17.4
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
wrapt==1.15.0
